refactor(eval): split api_runner by concern to clear module-health gate
